Rename CI_API files to clarify the API used (in case APIs are used for different purposes in the future)
