+++ cats/CI_api_query.py
@@ -2,7 +2,7 @@ import requests_cache
 from typing import Callable
 from datetime import datetime, timezone
 
-from .api_interface import API_interfaces
+from .CI_api_interface import API_interfaces
 
 def get_CI_forecast(postcode: str) -> list[list[tuple[datetime, int]]]:
     """
